hvf: Sign the code after installation
Before this change, the code signed during the build was installed
directly.

However, the signature gets invalidated because meson modifies the code
to fix dynamic library install names during the install process.

It also prevents meson to strip the code because the pre-signed file is
not marked as an executable (although it is somehow able to perform the
modification described above).

With this change, the unsigned code will be installed and modified by
meson first, and a script signs it later.

Message-Id: <20210225000614.46919-1-akihiko.odaki@gmail.com>
Message-Id: <20210226080526.651705-2-pbonzini@redhat.com>

+++ scripts/entitlement.sh
@@ -2,12 +2,23 @@
 #
 # Helper script for the build process to apply entitlements
 
+copy=:
+if [ "$1" = --install ]; then
+  shift
+  copy=false
+  cd "$MESON_INSTALL_DESTDIR_PREFIX"
+fi
+
 SRC="$1"
 DST="$2"
 ENTITLEMENT="$3"
 
-trap 'rm "$DST.tmp"' exit
-cp -af "$SRC" "$DST.tmp"
-codesign --entitlements "$ENTITLEMENT" --force -s - "$DST.tmp"
-mv "$DST.tmp" "$DST"
+if $copy; then
+  trap 'rm "$DST.tmp"' exit
+  cp -af "$SRC" "$DST.tmp"
+  SRC="$DST.tmp"
+fi
+
+codesign --entitlements "$ENTITLEMENT" --force -s - "$SRC"
+mv -f "$SRC" "$DST"
 trap '' exit
